Update patch set 8
Patch Set 8:

latest patch uses the flag approach that @Goktug pointed.

I have resolved the circular dependency with reflexion. BTW reflection was already used in this class for runing other server launchers.

In this way everything is pretty much simpler, since IDE vendors or launchers just have to add the -superDevMode flag and we dont have to maintain another launcher class.

I think in the future (or maybe now) we could turn on the superDevMode flag on by default, and users can revert to classic using the -nosuperDevMode option.

For testing with JUnitShell we need to do something else in order to fix urls, and junit.nocache.js but my tests are promising, so I could work on this in a separated patch.

I have changed javascript to be wrapped in a function, and the way to find a free port like @Brian suggested.

I have added a feature in the stub.nocache.js file marked as experimental until you guys give some feedback. What the javascript code does is to add a reload button at a corner to recompile the app, it also configures Ctrl-K for that if we preferred that option. This allows that we don't have to modify Shell-UI nor GPE in order to add a button to recompile the module, the user does not need to install bookmarkers.

I have updated the operation document and screenshots with the new stuff. Here you can see how the floating button in the corner looks like. I have added reviewers to shared list, if anyone wants please sendme a request.

https://docs.google.com/presentation/d/1T4RLzdtG84KkatyO7heTYLWlf1QqQwNZxrdQBuTrCuI

Patch-set: 8
Label: Verified=0