Update patch set 4
Patch Set 4:

It looks like computeTypeInstantiability() is supposed to return true if a type can be constructed, but doesn't consider the case where the type has a custom field serializer that has an instantiate() method. In that case it probably shouldn't be looking at the fields at all, because the custom field serializer will do all the work. So the audit check shouldn't be reached.

Does that make sense? If so, I think we should fix this in a separate patch, since it may have other consequences.

Patch-set: 4