Update patch set 2
Patch Set 2:

(1 comment)

> Goktug, does this change make any sense?
 > 
 > Can compiler optimize out equals call on strings?

IIRC the compiler will not optimize __equals because the parameter is referred twice in the body and inlining will not happen.

IMHO I prefer to make the compiler better in optimizing cases like this, rather than manually specialize it. I think manual specialization should be only used sparingly in situations where the gain in real world apps is clear (not micro benchmarks) and where it is really difficult to make the compiler handle them better.

Patch-set: 2