Update patch set 4
Patch Set 4:

Hi Brian,

Thanks for investigating; that is odd about j.s.Date. Without looking at the code, I agree with your computeTypeInstantiability diagnosis, and that we should patch it first.

Did you want to do that? That'd be great, if so. I can also take a look, but it might be a few days, and a few line numbers that look suspicious would be a good hint.

Patch-set: 4
Reviewer: Gerrit User 5025 <5025@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0