Update patch set 4
Patch Set 4:

I tried hacking it to work by adding this code in SerializableTypeOracleBuilder.checkDeclaredFields:

    if (typeInfo.isManuallyInstantiable()) {
      return true;
    }

Where isManuallyInstantiable is:

    public boolean isManuallyInstantiable() {
       return manualSerializer != null && 
         CustomFieldSerializerValidator.hasInstantiationMethod(
            manualSerializer, (JClassType) type);
    }

But lots of stuff breaks then. I suspect GWT-RPC has no principled way of figuring out which types to visit after reaching a type with a custom field serializer, and there is lots of code that just depends on current behavior (and I'm not sure what that is).

That there's another bug in this area:

https://code.google.com/p/google-web-toolkit/issues/detail?id=8728

It seems like if you're writing a custom field serializer, there should be some explicit way of declaring the types that it might serialize or deserialize. The problem is that sometimes this depends on on a type variable; imagine writing a custom field serializer for a collection.

Patch-set: 4