Update patch set 4
Patch Set 4:

I think I'm getting a handle on this.

There is this code at the end of checkDeclaredFields:

    boolean succeeded = allSucceeded || typeInfo.isManuallySerializable();
    if (succeeded) {
      typeInfo.setFieldSerializable();
    } 

What this means is that, for a class with a custom field serializer, GWT-RPC makes a best-effort attempt to find other types to visit, by scanning all the fields on the class in the normal way. It assumes that if it's possible to serialize a field, the custom field serializer will probably call writeObject() on that field, so GWT-RPC needs to handle it. On the other hand, if it's not serializable then the custom field serializer probably wouldn't try to call writeObject() on it, so we should just skip it and this doesn't make the type non-serializable.

GWT-RPC is only doing a "best-effort" field scan for custom field serializers, but enabling final fields is still a change in behavior since more types will be reachable. So I think it's right for the audit to report it, even though in the case of java.util.Date, it didn't catch a real bug.

So it seems like the safest thing is to make the JsDate field in java.util.Date transient (so the audit won't fail on it) and move on to audit more code.

Patch-set: 4