Super Dev Mode: support progress queries
A browser can find out whether the compiler is running by accessing
"/progress?_callback={functionName}". This is a JSONP callback that
returns a summary of the compiler's progress.

To start with, non-incremental compiles have two progress steps and
incremental compiles just have one. (Compiling or idle.)

Updated the bookmarklet's progress dialog to ping the server
once a second and print a dot if the compiler is still running.

(Also, fixed a thread safety issue: it was possible to start two
compiles at once if they were for different modules.)

Change-Id: I744327b2167ccdc768752daf7636fa9d9abc7817