Update patch set 6
Patch Set 6: Code-Review+2

At Google we are still using the old version of the DevMode entry point that was deleted from trunk a while back. It will need to be fixed, but it's not something you can do. I think the best thing is to commit this and if I need to do anything in trunk, I will follow up with another patch.

Patch-set: 6
Label: Code-Review=+2