Update patch set 3
Patch Set 3:

PTAL. I tweaked the classes in the preamble and now it should be completely safe (any class that has one of it methods or fields reachable from Class.createFor* is included).

It currently includes too much:
(1) GWT - due to GWT.isScript
(2) Impl - due to Object.hashCode
(3) UnloadSupport - due to Impl (and when unload support is enabled a bunch of other classes are included as well).
(4) Any JSOs that has a non empty $clinit - due to references to JavaScriptObject.

Patch-set: 3
Label: Verified=0