Update patch set 1
Patch Set 1:

While attempting to verify this change I discovered an interesting way that GWT-RPC can go wrong.

A very large GWT application happens to have a serializable class that contains a List<Exception>. This is probably a bad idea. But anyway, now all exceptions need to be serialized.

One possible exception is javax.validation.ConstraintViolationException.

GWT has a custom field serializer for this exception, because it has a final field. The type of that final field is:

java.util.Set<javax.validation.ConstraintViolation<?>>

Since it's final, the generator wasn't visiting this field before.

So, with this change, GWT-RPC sees the ? and decides that all subtypes of java.lang.Object have to be serialized. At the point things go downhill very quickly.

This is technically correct. It was possible before that GWT-RPC would blow up because someone tried to send a ConstraintViolation over the wire. But there's no way we can commit this change without mitigating this somehow.

I only discovered this because one of those subtypes had a bug, so the compile failed. But even if that were fixed, I would have to roll it back anyway since it would cause code boat.

And now you know why all changes to GWT-RPC are risky, particularly those that might cause it to crawl more types. A lot of things are working by coincidence.

Patch-set: 1
Label: Verified=0