Update patch set 1
Patch Set 1:

Wow. Sounds like fun...

So, this is probably a terribly naive question, but:

> GWT-RPC sees the ? and decides that all subtypes of java.lang.Object have to be serialized.

Given "serializing everything" is basically never the right thing to do, I'm surprised GWT-RPC doesn't short circuit on the wildcard (or even short circuit on seeing java.lang.Object).

However, since it doesn't short circuit, I guess I should assume it does this on purpose (probably for a good reason?), and that we can't just tell it to not do that...

Patch-set: 1
Label: Code-Review=0