Enables the -XcompilePerFile flag for the regular Compiler.
Turning on this option also coerces several incompatible options to
compatible values.

The impact of -XcompilePerFile in a single direct invocation of
Compiler.java is limited to usage of the new per-class linking process.
To get the per-class caching and faster recompiles will at least
initially require use of SuperDevMode to manage repeated compiles within
a persistent JVM process.

Change-Id: I4a4747b5354b29df2f36e9e27bdf65f988bd67f7