Do not record location for inlined code in JsStackEmulator.
Stack trace deobfuscation and reporting mix information about
the function appearing in the stack trace and the instrumented
line number and this to do not correspond to the location of
the actual code.

Now the stack frame for the inlined function will not appear in the
stack in any way; a better solution would be to artificially add an
"inlined" stack frame so that both the call location and the
inlined callee location appear in the stack frame.

Change-Id: I93b821a37b8447dd28254e8da02d90bda1e56db2