Consider live any array that is passed or returned from JSNI.
Arrays created directly in JSNI were allowed in some circumstances
as it is used in the SDK for convenience.

This patch makes live any array type that is returned from JSNI or
passes from JSNI to Java.

Arrays created in this way will mostly work. If arrays created this
way are explicitly cast or checked via instanceof an exception might
occur.

Change-Id: I5482165a1e504de97a54e5514ec5be90dc84e3ea