Update patch set 5
Patch Set 5:

> (6 comments)
 > 
 > I'm a little unclear what this change is for. Is it to support
 > compiling multiple permutations?

When you first compile the app, if user.agent is unknown it compiles all permutations, this happens with -superDevMode parameter if useragent is not guess by the script.
Also if the user is testing with two browsers at the same time, he compiles for each browser and there are two permutations in the same folder.
Both cases causes source maps not being served raising a server runtimeException in ModuleState.findSourceMap with the message: 
"Multiple fragment 0 sourcemaps found. Too many permutations."

Patch-set: 5