Update patch set 6
Patch Set 6:

> I am not sure if it is worth but it would be cool to check in
 > jsperf how does this implementation compare for small, medium and
 > large json texts?

I ran it for a relatively small json and a relatively large one.
http://jsperf.com/safe-json-evaluation

In Chrome the usage of regular expression is insanely fast for the small json.
Both in Chrome and Firefox the regular expression is insanely slow for  the large json.

Using JSON.parse() is quite slow when the json is not safe. But it seems to be the best for large json files.

The Finite State Machine seems to be always in the middle. Maybe it could be improved by using int instead of enums.

Patch-set: 6