Update patch set 1
Patch Set 1:

> FYI, https://gwt-review.googlesource.com/#/c/1680/

@Andrei, I din't released that patch, I'm wondering why we have patches so long in gerrit. 

Anyway, this is almost the same patch, I prefer to test JS implementation running tests with the  -prod flag instead (see comment in your review).

We can do either, abandon your old change in favor of this, or viceversa. But i need any of them going to master in order to have tests running with ant and passing in both Dev and JS.
I'm preparing a patch to run Collections and Json tests in JRE so as it's reliable to use elemental json in server side.

Patch-set: 1
Label: Verified=0