Update patch set 2
Patch Set 2: Code-Review+1

I think this doesn't warrant a test.

Wr.t. new number; 100ms is clearly too long will make any webpage with incremental commands to act sluggish and have around 10 fps.
On the other hand, we can switch to 16 (instead of 8) which is more conservative change and will make it possible to keep the fps around 60.
Not so far in the future, we can make this easily configurable via [1] but 16ms still sounds like a good default...

[1] https://docs.google.com/a/google.com/document/d/1LEP0H-6z__Mog0QTt-NAFpllcOGu7X5gr7KR32BIT4Y/edit

Patch-set: 2
Label: Code-Review=+1