Update patch set 2
Patch Set 2:

I think 8ms is the correct number.  This is the minimum amount of time the scheduler will run.  At 16ms any task whose running time isn't a factor of 16 will drop the framerate below 60.  

So the issue is that a larger TIME_SLICE means a shorter time to complete.  I think a good estimate of the worst case of the time to complete difference is to calculate it using the TIME_SLICE and FLUSHER_DELAY values.  Where a task runs for the TIME_SLICE the browser regains control during the FLUSHER_DELAY and then the task runs again.  

FLUSHER_DELAY is 1ms. In the worst case a task that takes a 16ms TIME_SLICE 1000ms to complete will take an 8ms TIME_SLICE 1062ms to complete.  So not much difference. The major factor making a difference to how long a task will run will be the CPU speed not the TIME_SLICE.

Anyway I have an in development app affected by this issue.  I'll switch it over to 16ms and report back on what happens in a few hours.

+1 Agree this should be a tunable property.

Patch-set: 2