Update patch set 2
Patch Set 2:

Can't we just remove the while loop in question and just keep the for loop inside that while? That would mean we go through the JsArray of tasks, executes each task once and then give the browser FLUSHER_DELAY time to breathe.

Currently the while loop ensures 100ms of execution time as long as tasks are available. However when dropping it down to 16ms or 8ms I am not sure if it is actually worth it to keep that behavior.

If we would remove it then the browser can only lag if <amount of tasks> * <average task execution time> becomes too long. But the developer can control both of them.

But maybe I am missing something?

Patch-set: 2
Label: Verified=0