Update patch set 2
Patch Set 2:

I think getting rid of the while loop will probably work but I'm not certain.  I think if each instance of the repeating task is very short (less than 2 or 3ms and especially if the task takes only a fraction of a ms) then it might cause the total execution time to expand quite rapidly.

I don't think my bogo sort demo will be able to demonstrate the execution time expansion.  I'm going to need to rewrite the app that I detected the problem on to tell you if getting rid of the while loop is a good idea so I'll let you know in a few days.

I can't share my app with you but basically what I'm doing is showing the user a linear series of css animations each of which plays for about 10 seconds.  The animations need to be calculated at runtime so every 20 animations or so I refill my animation buffer using a scheduled task.

The basic unit for calculating my animation is to write text to the dom and then get an element's height which I guess might often take less than a ms.  But I need to do this multiple times (150 maybe is a good estimate for the worst case) for each animation.  It breaks my app if animations aren't perfectly smooth but users can also skip animations at will by tapping the screen so each animation has to be calculated within about 100ms or a user might skip past my buffer.

Anyway I'll test removing the while loop entirely and let you know asap what it does to my execution times.

Patch-set: 2