Update patch set 2
Patch Set 2:

There is no "1 ms" rescheduling, FLUSHER_DELAY is a big lie ;)
All browsers do clamping and will schedule at 4ms to 10 ms that is why you are seeing a giant jump from 2 secs to 3 secs when you change slice 16 to 8. That is a big impact so that's why I recommended 16ms as it is a more conservative move.

Also to clarify, I said it will *be possible* to keep the FPS around 60, I didn't say "it will make it 60 fps". With 16ms slice, you can keep the tasks small enough that you can avoid going beyond 16.67ms. Perhaps we can make it 15ms to provide more room but in general I think that's good enough.

If you can prove me otherwise with an example that your UI feels sluggish, let's talk about the alternatives.

Patch-set: 2