Make polymorphic dispatch work in v8.
Classes that ONLY override methods from parent classes might not
be optimized in v8, due to a problem with the way hidden classes are
built.

V8 requires prototypes of different classes to differ in the set of
properties in order to perform some optimizations. This patch makes
every prototype have a unique field for that purpose.

This optimization was attempted before with a different implementation
at commit 340d85f5b30010bd4b6e358457711b08f2be2d86; but was rolled
back due to performance regression in IE.

Change-Id: I813a9e983d2abceddfda7484136d845d7a7881dd