Update patch set 1
Patch Set 1:

I'd like to avoid a big rewrite for this patch, but I'll comment on the larger design issue:

File and directory watching is rather complex and finicky. The low-level filesystem API's for this are different for each OS and support for them varies by filesystem, so testing this code is sort of like making sure code runs in each browser. (I've written code like this before but it was internal so I could at least assume Linux.) We also need to handle jar files. The Java filesystem API's don't really cover up the different behavior. 

Also, some filesystems are very large and scanning them is slow. This is why there have been lots of problems with IDEA and Eclipse with Google's internal filesystems over the years.

So, I think we shouldn't look outside the directories that the GWT compiler knows about. Also, we should avoid file-watching API's and stick with portable file API's that are available everywhere.

I think a good approach would be to make a list of every input file that the GWT compiler sees, and remember each file's path, timestamp, and size since these are cheap to query. Then just do an equality comparison with the same list from the previous compile. For our purposes, it doesn't matter whether a file's timestamp is older or newer, just that it's different. (Timestamps can go backwards sometimes, for example when reverting a zip file change.)

Luckily for this particular patch, the consequences for getting the check wrong are not that bad; the user can touch any source file to force a recompile.

Patch-set: 1