Update patch set 2
Patch Set 2:

> File and directory watching is rather complex and finicky. The
 > low-level filesystem API's for this are different for each OS and
 > support for them varies by filesystem, so testing this code is sort
 > of like making sure code runs in each browser. (I've written code
 > like this before but it was internal so I could at least assume
 > Linux.) We also need to handle jar files. The Java filesystem API's
 > don't really cover up the different behavior.
 > 

Is Java's java.nio.file.WatchService broken or doesn't cover our use case?

 > Also, some filesystems are very large and scanning them is slow.
 > This is why there have been lots of problems with IDEA and Eclipse
 > with Google's internal filesystems over the years.

I thought there was another reason for that (i.e. complex incl. excl. patterns to mimic blaze globs) but I guess we can try and see if it is very slow or not.

 > 
 > So, I think we shouldn't look outside the directories that the GWT
 > compiler knows about. 

I thought superdev mode has already an understanding of some kind of source root concept but anyway, this is not a big deal for me especially as you pointed the consequences are not that bad.

 > Luckily for this particular patch, the consequences for getting the
 > check wrong are not that bad; the user can touch any source file to
 > force a recompile.

You are right. User will see that superdev didn't see any changes so that's not a big deal.

Patch-set: 2