Update patch set 2
Patch Set 2:

"The implementation that observes events from the file system is intended to map directly on to the native file event notification facility where available, or to use a primitive mechanism, such as polling, when a native facility is not available. Consequently, many of the details on how events are detected, their timeliness, and whether their ordering is preserved are highly implementation specific."

http://docs.oracle.com/javase/7/docs/api/java/nio/file/WatchService.html

In other words, they don't guarantee that it will perform well. There are situations where this would be hard to do, such as an NFS-mounted filesystem.

Patch-set: 2