Update patch set 2
Patch Set 2: Code-Review+1

> "The implementation that observes events from the file system is
 > intended to map directly on to the native file event notification
 > facility where available, or to use a primitive mechanism, such as
 > polling, when a native facility is not available. Consequently,
 > many of the details on how events are detected, their timeliness,
 > and whether their ordering is preserved are highly implementation
 > specific."
 > 
 > http://docs.oracle.com/javase/7/docs/api/java/nio/file/WatchService.html
 > 
 > In other words, they don't guarantee that it will perform well.
 > There are situations where this would be hard to do, such as an
 > NFS-mounted filesystem.

For the platforms we most care (linux & windows) and for common use cases this should work fine and should not require polling; if not we can simply use jnotify instead (which claims it doesn't poll in Mac, Linux and Windows). Where the performance degrades, we could simply give an option to disable it.

Anyway, as it is a small patch I don't want to block it (hence +1) but I still think the proper solution is adding a watcher and later as next step add automatic trigger of the compilation.

Patch-set: 2
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1