Update patch set 11
Patch Set 11:

> Repackaging would be nice.
 
Well, i strongly disagree with this approach, as i said on another change [1]. That feels, like very wrong thing to do for me. In fact that's the very last option you have and by doing it you are actually saying: "I failed to do it the right way, I will repackage it and hide and burrow all it inside my own artifact, so that this doesn't smell." But it still does, and this wouldn't reduce the pain of GWT users.

 > For Super Dev Mode, it's somewhat tempting to switch off of Jetty
 > to something like NanoHttpd (and fork it and do what we like).

Agreed.

 [1] https://gwt-review.googlesource.com/#/c/7857/13//COMMIT_MSG

Patch-set: 11