Update patch set 11
Patch Set 11:

> > Repackaging would be nice.
 > 
 > Well, i strongly disagree with this approach, as i said on another
 > change [1]. That feels, like very wrong thing to do for me. In fact
 > that's the very last option you have and by doing it you are
 > actually saying: "I failed to do it the right way, I will repackage
 > it and hide and burrow all it inside my own artifact, so that this
 > doesn't smell." 

What would "the right way" be?
Updating to Jetty 9.2 will break all those in your situation but using Jetty 8 (or any other version of Jetty that conflicts with the version we'd pick).
Should we rather embed the Jetty JARs within the gwt-dev.jar and use a specific ClassLoader? (a bit similar to how Gerrit works when launched with "java -jar gerrit.war", and how One-JAR works, and how JDT apparently works – we're extracting a JAR from the JDT JAR, they seem to be using this to "hide" JDT internals that we actually use)

Or we could just move Jetty to a com.google.gwt.thirdparty.org.eclipse.jetty package tree using JarJar.

 > But it still does, and this wouldn't reduce the
 > pain of GWT users.
 
Why?
Think of a repackaged Jetty as if we had forked it or written a web server ourselves, it's in GWT's packages so it won't conflict with anything that already exists, and it's internal so you're not supposed to depend on it; that way, it cannot conflict with anything.

We need a specific version of Jetty due to HtmlUnit (and forward/backward-incompatibilities with Jetty itself), just like we need a specific version of JDT because we use internal APIs. We could say "don't use any other version or it will break", but that's actually what we're implying today and it hurts (Gerrit wrt Jetty, others wrt JDT); or we could "hide" them as they're implementation details that we don't want to leak to the outside. The easiest way is to just rename the packages (this is what Guice uses for its some of its dependencies; now that ASM and Guava provide forward-compatibility guarantees they're thinking about reverting this, but it was necessary with earlier versions of those tools; this is also what the JDK does for Xerces, and I'm sure there are many other examples).
Splitting gwt-dev so you could avoid the dependency on HtmlUnit wouldn't help us, because people using the dependency would then run into the conflict; it just happens that Gerrit doesn't use GWTTestCase so you don't care.
 
 > > For Super Dev Mode, it's somewhat tempting to switch off of Jetty
 > > to something like NanoHttpd (and fork it and do what we like).
 > 
 > Agreed.
 
How would that be different?

Repackaging Jetty is not much different from forking NanoHttpd; or if you fork without changing the package then you fall into the same issues than today with Jetty: someone using NanoHttpd in his project could stumble on incompatibilities/conflicts between the version he uses and the one we use in GWT; except maybe NanoHttpd has a more stable/frozen public API. People using Undertow, Netty or Tomcat in their applications don't have conflict issues with the Jetty we bundle/use, so switching to anything other that is not labelled "GWT specific" (forked or repackaged server, whichever it is) is likely to create conflicts in other projects.

(if your agreement was on using a lighter-weight HTTP implementation, then feel free to ignore my comment above)

Patch-set: 11