Update patch set 11
Patch Set 11:

Just a thought. Instead of updating it, another method that could provide the latest jetty easily could be providing a app engine launcher like the AppEngineLauncher which could be run in the -server argument. If a library had launcher(s) for different jetty versions, could be used and piped in with the -server. Currently the DevMode is botting jetty which has a static version, but that can be easily overridden. With a -server override the IDEs don't even need to be updated to run a newer web server. Although I've been leaning more to the WTP facets configuration which extracts the logic into the editor to run the web servers or servers that hosts the code. So anyway, if a launcher was created for the newer jetty version, it could be imported in easily a maven dependencies, or manually downloaded and added to the classpath.

Patch-set: 11
Reviewer: Gerrit User 5625 <5625@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0