Update patch set 11
Patch Set 11:

> Just a thought. Instead of updating it, another method that could
 > provide the latest jetty easily could be providing a app engine
 > launcher like the AppEngineLauncher which could be run in the
 > -server argument.

This is what Gerrit does (did?), but AppEngine spawns a distinct process. That means that the launcher would need a mean to gather the classpath for the spawned process (for AppEngine, this is done by providing the path to the SDK). If you can do that, you could just as well create a ClassLoader for those JARs and launch your server in that ClassLoader, which you could easily isolate from the gwt-dev classpath. The problem is that it's not that easy to know which JARs or folders to use.

So Gerrit had been, for now, stripping Jetty from gwt-dev to make tusre it won't conflict with their version, and using their own ServerContainerLauncher with a more recent Jetty.
The problem is that this cannot be done (easily) for SDM, as SDM is itself a webapp and it directly uses the Jetty APIs, which have changed in incompatible ways in Jetty 9 compared to Jetty 8. This is the crux of the problem, and what David is fighting with.

Patch-set: 11