Introduces specialization for String.equals(String).
In order to that, I needed make static eval work with method
overloads so that when things optimized to specialize method
(either through inlining of equals(Object) or @SpecializeMethod)
we can still find a matching method in String class.

However, this intorduced another problem. If we introduce an
overload equals, the JDT is generating calls with wrong method
signature if the type passed is a String in the compile time and
that was breaking devmode.

To fix that I updated @SpecializeMethod to work with private
methods so that I can make the equals(String) private which will
prevent JDT to match it from external calls.

Also fixed a bug where we were not correctly extracting the
annotation values from JDT.

Change-Id: If207180351f72fd5893b8b79373c5b57f6a0e638