Update patch set 3
Patch Set 3:

(6 comments)

This change baiscally keeps two copies around of the conditionalValues.

One that was original fed by the parse of the modules and one that can be altered by SDM without it running into problems asking the next time if a value is valid or not.

Patch-set: 3
Label: Verified=0