Super Dev Mode: stop reparsing the ModuleDef on every refresh (take 2)
Starts trusting ModuleDefLoader.loadFromResources() to only reparse the
ModuleDef if it detects that any of the .gwt.xml files have changed.

Since we are reusing the ModuleDef, we need to avoid changing anything
that would affect the next compile. In particular, isAllowedValue()
shouldn't be affected by temporary binding property overrides.

So, distinguish between allowed values and generated values in the
BindingProperty objects. We can temporarily override the generated values
with affecting the allowed values. In a normal compile, generated values
are the same as allowed values but in Super Dev Mode, they can be
a subset.

Did some refactoring in BindingProperties to make this clear.

Change-Id: I00dac15b972717a9fc0fb93f04274f73c5509591
Review-Link: https://gwt-review.googlesource.com/#/c/8778/