Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > Hi tomas, could you rebase this patch? I would like to add this
 > task in jenkins.

Done. But did you know you can actually even do it yourself?
Checkout the change (click on the "Download" dropdown on the top-right to pick the right command), rebase or amend it as needed (with the "best practice" of leaving the original "Author" of the commit) and push again to refs/for/master, and it'll update the review. Gerrit is wonderful! ;-)

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0