Update patch set 3
Patch Set 3:

> Do you have any idea whether precompiled modules speed-up later
 > targets in the build (such as tests, particularly in prod mode) ?

Yes I know, I run all test which we run in jenkins and the time was almost 
identical. What it was worst was sample compilation, so I left them after
running precompilation.
But anyway you are right, it's safer to leave things at they were and simply
add a flag to disable precompile.modules like we do in other places.

I'll send another patch set with this modification so as it looks better
for you.

Patch-set: 3
Label: Verified=0