Update patch set 4
Patch Set 4:

You've switched to building distinct artifacts depending on a flag, which is better, but still not ideal. Now, maybe our "incremental" build is far from ideal and it doesn't really matter, I don't know: it would be a problem if building the project once with the flag set to false and next with the flag set to true would leave the JAR untouched (and thus without the gwtar files) because the input files have not changed. If we unconditionally recreate the JARs, then I'm OK with adding that flag, otherwise I think we could do better.

Patch-set: 4
Label: Verified=0