Update patch set 4
Patch Set 4:

It works fine in these cases. So if first build with the flag and second without it and without cleaning, the artifact is updated with the .gwtar stuff. And if you first build without the flag the artifact always contains the precompiled stuff.

The proposal of this patch is for two cases:

- speedup builds in CI (presubmit)
- easy to run individual tests and find which commit is the responsible of a failure:

  $ git checkout <commit_id>
  $ ant clean user -Dtarget=test  -Dprecompile.disable=true \
  -Dtest.dev.disable=true \
  -Dtest.web.htmlunit.disable=true \
  -Dtest.nometa.htmlunit.disable=true \
  -Dtest.emma.htmlunit.disable=true \
  -Dtest.nongwt.disable=true \
  -Dtest.coverage.htmlunit.disable=true \
  -Dgwt.junit.testcase.includes=**/AutoBeanCodexJreTest.class

Patch-set: 4
Label: Verified=0