Super Dev Mode: put back old sourcemap URL and refactor
It turns out we were using the old sourcemap URL [1] in some tests. I think
it's worth keeping since it's nice to be able to download the sourcemap
without knowing the strong name. (Of course this only works with one
permutation.

I also refactored a bit to distinguish between the names of sourcemap files
on disk versus the names of sourcemaps in URL's. These don't need to be
the same. (And I changed the URL slightly so they aren't the same.)

[1] at /sourcemaps/[module]/gwtSourceMap.json

Change-Id: I366b9b13b0026d922428ae3f1ae56abad9d07059