Honor saveSources in per file compiles.
Save sources relied on a non implicit communication from the
compiler where the saveSource flag is processed to the linker phase
where the save source process is done.

Sources were saved in the linker phase if any emitted artifact with
visibliity SOURCE is present in the set of artifacts. The compiler
phase removes all SOURCE artifacts if saveSources is not set.

ASIDE: Checking whether there are generated source artifacts could
lead to an error when there are no generated sources, but saveSources
is set.

This patch makes the proper option available to the linker phase.

TODO: cleanup the saveSource implementation by letting the options
arrive at the right place.

Change-Id: Ia4f000c011c6cac48ecc842e423ba55c6bd4ebf5