Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

I defer to John's expertise on this. We need a solution for shared code where people are calling toUpperCase(Locale.ENGLISH), and where Turkish locale is busting basic things like object serialization or Enum parsing.

Long term I hope we can find a way to support even more of the JDK Locale stuff and synchronize somehow with GWT's notion of a Locale. This will seem to depend on waiting for browsers to get smarter (there's a new Encoding API, and Intl API, but it's not enough), or dumping in a massive polyfill.

Short term, landing something that only works for a very strict subset of use cases and erroring if used in other ways seems acceptable.

Patch-set: 1
Label: Code-Review=+2