Update patch set 12
Patch Set 12:

Manolo: it turns out that the new file-per-compile option that John is working on is incompatible with multiple permutations. See [1]. I'm wondering if that makes this patch obsolete. 

Is there some particular reason we need to support multiple permutations? (I forgot why you originally wanted this patch.)

[1] https://gwt-review.googlesource.com/#/c/9070/

Patch-set: 12