Update patch set 1
Patch Set 1:

There are two parts to this, why have an embedding API at all and why add this particular monitoring API?

Build tools (like Blaze and Buck) and IDE's need to start Super Dev Mode, and IDE's need to this as well. We want to encourage these tools to add better Super Dev Mode support, so it's useful to think about how to support them officially. We want to avoid the situation where they implement integration but it breaks with a new GWT release. However, as you say, maintaining a public API is a burden and that's why most classes are not public; I'm deliberately keeping the API surface fairly low.

The simplest integration is to call CodeServer.main(String[] args) but this calls System.exit() if there's an error which isn't so good if the tool wants to keep running. I added main(Options) and start(Options). The start() method is particularly useful because it avoids calling System.exit(). I use this entry point for starting the code server in tests. main(Options) is an intermediate version for when you want to supply Options but it's okay call System.exit().

A second question is about whether we want a public API to monitor the code server while it's running. Actually, it looks like there will be two ways to do this: polling a URL such as /policy and getting JSON back, which is useful in a web browser. There is also this event interface which is synchronous and type safe, so it will be more convenient to use from Java. Why have two?

The main reason I'm adding this API is for better Silo integration. But I thought it was worth cleaing it up and making it public because I think other embedders would also want to monitor Silo's progress and display progress information while it's running, much like a JUnit runner shows test progress in an IDE. This is admittedly speculative but I think the risk is fairly low since JobEvent is a simple record-like class.

Patch-set: 1