Update patch set 1
Patch Set 1:

You are the one who knows and thinks most about his so if you feel strongly about it go ahead but I would personally wait until Blaze/Buck or an IDE to approach to us for an integration so that we can better understand the use cases, then evaluate and iterate the API based on that instead of coming up with one today and make breaking changes later.

Similarly, for the java monitoring API needed by Silo, at least for today, it looks like it is pretty unique to our needs. So again I would avoid committing to a public API and maintain it until there are other more compelling use cases. In the meantime we can use some kind of private integration point for our internal needs and freely modify it based on what we need.

Patch-set: 1
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0